aws cloudstorage bucket refactor
