+++ controllers/bsl.go
@@ -134,9 +134,26 @@ func (r *DPAReconciler) ReconcileBackupStorageLocations(log logr.Logger) (bool, 
 				}
 				bsl.Spec.Credential = bslSpec.CloudStorage.Credential
 				bsl.Spec.Default = bslSpec.CloudStorage.Default
-				bsl.Spec.Provider = AWSProvider
 				bsl.Spec.ObjectStorage = &velerov1.ObjectStorageLocation{
 					Bucket: bucket.Spec.Name,
+				}
+				switch bucket.Spec.Provider {
+				case oadpv1alpha1.AWSBucketProvider:
+					{
+						bsl.Spec.Provider = AWSProvider
+					}
+				case oadpv1alpha1.AzureBucketProvider:
+					{
+						panic("Azure provider not yet supported")
+					}
+				case oadpv1alpha1.GCPBucketProvider:
+					{
+						panic("GCP provider not yet supported")
+					}
+				default:
+					{
+						return fmt.Errorf("invalid provider")
+					}
 				}
 			}
 			return nil
